Refactor profile header out of patient profile index
