+++ src/Language/Locales/en/translation.json
@@ -912,6 +912,6 @@
         "thanks": "Thank you for your input",
         "recorded": "Your response has been recorded",
         "reachOut": "Not all of your contacts have been tested yet, please reach out to your treatment assistant for more information and recommendations.",
-        "completed": "Great job ensuring the health of those you live with."
+        "completed": "Great job ensuring the health of those with whom you live."
     }
 }
